Move generatePathWithSearchParams to paramsSlice.
The generatePathWithSearchParams function should not include query parameters
whose values are the default values for those parameters. Those defaults are
defined in paramsSlice, so it makes sense for the function to be defined there.
The generatePathWithSearchParams function should be used when constructing any
other path involving parameters, and this commit updates a few places where
paths were being created in some other way.
